use expandCopies function
